feat: implement product entity, repository, and exceptions; update migration and model definitions

+++ migrations/20260207145727-create-products-table.js
@@ -18,7 +18,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.DECIMAL(10, 2),
+        type: Sequelize.DOUBLE,
         allowNull: false,
       },
       stock: {
